Change `ForgeDirection` side serialization to `byte`

+++ src/main/java/com/glodblock/github/network/SPacketLevelTerminalUpdate.java
@@ -269,7 +269,7 @@ public class SPacketLevelTerminalUpdate implements IMessage {
             buf.writeInt(y);
             buf.writeInt(z);
             buf.writeInt(dim);
-            buf.writeInt(side);
+            buf.writeByte(side);
             buf.writeInt(rows);
             buf.writeInt(rowSize);
 
@@ -302,7 +302,7 @@ public class SPacketLevelTerminalUpdate implements IMessage {
             this.y = buf.readInt();
             this.z = buf.readInt();
             this.dim = buf.readInt();
-            this.side = buf.readInt();
+            this.side = buf.readByte();
             this.rows = buf.readInt();
             this.rowSize = buf.readInt();
             int payloadSize = buf.readInt();
